keep refactoring into separate client crates - now compiles
