+++ client/src/propose/mod.rs
@@ -1,0 +1,3 @@
+pub mod bpf;
+pub mod multisig;
+pub mod token;
